Time: 104 ms (62.12%) | Memory: 141.8 MB (93.75%) - LeetSync

+++ 2129-number-of-pairs-of-interchangeable-rectangles/number-of-pairs-of-interchangeable-rectangles.cpp
@@ -7,7 +7,7 @@ public:
             double r1=(double)rec[i][0]/rec[i][1];
             mp[r1]++;
         }
-        for(pair<double,int> p:mp){
+        for(pair<double,long long> p:mp){
             long long n=p.second;
             cnt+=(n*(n-1)/2);
         }
